+++ src/hip/hip_create_allocate.cpp
@@ -49,12 +49,12 @@ using std::string;
 //   return data;
 // }
 
-long get_malloc_size_host(mblas_data_type type, long x, long y, int batch,
+long long get_malloc_size_host(mblas_data_type type, long x, long y, int batch,
                           long long stride) {
   int typesize = type_call_host<sizeofCUDT>(type);
   long packing_count = type.get_packing_count();
-  long base = x * y;
-  long total_elements;
+  long long base = x * y;
+  long long total_elements;
   if (batch > 1 && stride > base) {
     total_elements = stride * (batch - 1) + base;
   } else {
@@ -63,12 +63,12 @@ long get_malloc_size_host(mblas_data_type type, long x, long y, int batch,
   return total_elements * typesize * packing_count;
 }
 
-long get_malloc_size_dev(mblas_data_type type, long x, long y, int batch,
+long long get_malloc_size_dev(mblas_data_type type, long x, long y, int batch,
                          long long stride) {
   int typesize = type_call_dev<sizeofCUDT>(type);
   long packing_count = type.get_packing_count();
-  long base = x * y;
-  long total_elements;
+  long long base = x * y;
+  long long total_elements;
   if (batch > 1 && stride > base) {
     total_elements = stride * (batch - 1) + base;
   } else {
